Initial commit of snp_fetch
This is the initial commit of the snp_fetch script.  The script represents the state-of-play for it's intended purpose.  The script reads a PGS file, looks up rsIDs based on chromosome/base_position pairs found in the PGS file.  It then writes a new PGS file with the same contents as the original PGS but with the rsID column added.  If no rsID can be matched, then the field will be empty.