Add rsID duplication tool
rsid_cross_check reads a number of PGS files which have rsIDs in them and records which rsIDs occur in which files.  Then the list is transposed (in essence) to report on rsIDs that occur in multiple files.
Currently there is no specific detection of duplicates within a file.  However, if a file is list twice on the command line it will only be reported once (though the count will be wrong.