Add tool to report used rsIDs
The used_rsids.py tool takes a collection of base files (which contain rsIDs in some column) and a collection of removed files which records which rsIDs were not used during the proecessing.  The tool then does set subtraction on the two sets of rsIDs and produces a csv style output containing all the rsIDs that were used.

The tool allows the column number to be specified for the two different file types, but defaults to 0 if left unspecified.  It is possible to specify both base file column and the removed file column, either alone or not specify either.

The tool allows the specification of an output file or will use standard output if no output file is given.

Additionally this change adds a couple of test files for this tool.