Allow rsid_cross_check to read plain txt files
Allow rsid_cross_check to read in plain text files where the rsIDs are listed one per line in the first column.

This change also adds in the intended ability to limit the output to only rsIDs that appear at least N times.  The code was there to do this, but there was no command line argument for it.