Catch blank lines and avoid exceptio
Check to see if line read is blank (a condition which would lead to words[0] getting a list index out of range) and simply warn that it has happened, skip it and carry on.

This prevents the code from crashing but provides useful notice since a blank line was probably not intended.